+++ gdynamicDigitization/gParticleCounterDigitization.cc
@@ -2,8 +2,8 @@
 
 
 bool GParticleCounterDigitization::defineReadoutSpecsImpl() {
-	float timeWindow    = 10;                  // electronic readout time-window of the detector
-	float gridStartTime = 0;                   // defines the window grid
+	double timeWindow    = 10;                  // electronic readout time-window of the detector
+	double gridStartTime = 0;                   // defines the window grid
 	auto  hitBitSet     = HitBitSet("000000"); // defines what information to be stored in the hit
 
 	readoutSpecs = new GReadoutSpecs(timeWindow, gridStartTime, hitBitSet, digi_logger);
