+++ gfactory/examples/FordFactory.h
@@ -7,7 +7,7 @@ class Ford : public Car
 public:
 	Ford();
 	void go();
-	float fordVar[2]{};
+	double fordVar[2]{};
 };
 
 extern "C" Car* CarFactory(void) {
