+++ gfactory/examples/TeslaFactory.h
@@ -7,7 +7,7 @@ class Tesla : public Car
 public:
 	Tesla();
 	void go();
-	float autopilot[2][2];
+	double autopilot[2][2];
 };
 
 extern "C" Car* CarFactory(void) {
